move DagOperations bfTraverse to shared dagOps
