+++ casper/src/main/scala/coop/rchain/casper/SafetyOracle.scala
@@ -11,12 +11,12 @@ import coop.rchain.blockstorage.dag.BlockDagRepresentation
 import coop.rchain.blockstorage.syntax._
 import coop.rchain.casper.protocol.Justification
 import coop.rchain.casper.util.ProtoUtil._
-import coop.rchain.casper.util.{Clique, DagOperations, ProtoUtil}
+import coop.rchain.casper.util.{Clique, ProtoUtil}
 import coop.rchain.models.BlockMetadata
 import coop.rchain.metrics.{Metrics, Span}
 import coop.rchain.models.BlockHash.BlockHash
 import coop.rchain.models.Validator.Validator
-import coop.rchain.shared.{Log, StreamT}
+import coop.rchain.shared.{DagOps, Log, StreamT}
 
 /*
  * Implementation inspired by Ethereum's CBC casper simulator's clique oracle implementation.
@@ -165,7 +165,7 @@ sealed abstract class SafetyOracleInstances {
               val creatorJustificationOrGenesis = block.justifications
                 .find(_.validator == block.sender)
                 .fold(block.blockHash)(_.latestBlockHash)
-              DagOperations
+              DagOps
                 .bfTraverseF[F, BlockHash](List(latestByValidatorHash)) { blockHash =>
                   ProtoUtil.getCreatorJustificationAsListUntilGoalInMemory(
                     blockDag,
