+++ casper/src/main/scala/coop/rchain/casper/Validate.scala
@@ -11,7 +11,7 @@ import coop.rchain.blockstorage.syntax._
 import coop.rchain.casper.protocol.{ApprovedBlock, BlockMessage, Justification}
 import coop.rchain.casper.util.ProtoUtil.bonds
 import coop.rchain.casper.util.rholang.RuntimeManager
-import coop.rchain.casper.util.{DagOperations, ProtoUtil}
+import coop.rchain.casper.util.ProtoUtil
 import coop.rchain.crypto.codec.Base16
 import coop.rchain.crypto.hash.Blake2b256
 import coop.rchain.crypto.signatures.Secp256k1
@@ -232,7 +232,7 @@ object Validate {
       maxBlockNumber      = ProtoUtil.maxBlockNumberMetadata(initParents)
       earliestBlockNumber = maxBlockNumber + 1 - expirationThreshold
       _                   <- Span[F].mark("before-repeat-deploy-duplicate-block")
-      maybeDuplicatedBlockMetadata <- DagOperations
+      maybeDuplicatedBlockMetadata <- DagOps
                                        .bfTraverseF[F, BlockMetadata](initParents)(
                                          b =>
                                            ProtoUtil
